Made requested changes.
