+++ open_instruct/test_grpo_fast.py
@@ -799,25 +799,5 @@ class TestStreamingAccumulation(TestGrpoFastBase):
         self.assertEqual(len(pending_queries_map), 0)
 
 
-class TestUtilityFunctions(unittest.TestCase):
-    """Test utility functions in grpo_fast module."""
-
-    @parameterized.expand(
-        [
-            ("basic_repeat", ["a", "b", "c"], 3, ["a", "a", "a", "b", "b", "b", "c", "c", "c"]),
-            ("repeat_once", ["a", "b", "c"], 1, ["a", "b", "c"]),
-            ("repeat_zero", ["a", "b", "c"], 0, []),
-            ("empty_sequence", [], 3, []),
-            ("integers", [1, 2, 3], 2, [1, 1, 2, 2, 3, 3]),
-            ("mixed_types", ["a", 1, None, True], 2, ["a", "a", 1, 1, None, None, True, True]),
-            ("single_element", ["x"], 5, ["x", "x", "x", "x", "x"]),
-        ]
-    )
-    def test_repeat_each(self, name, sequence, k, expected):
-        """Test the repeat_each function with various inputs."""
-        result = grpo_fast.repeat_each(sequence, k)
-        self.assertEqual(result, expected)
-
-
 if __name__ == "__main__":
     unittest.main()
