+++ open_instruct/utils.py
@@ -70,6 +70,11 @@ logger = get_logger(__name__)
 DataClassType = NewType("DataClassType", Any)
 
 
+def repeat_each(seq, k):
+    """Repeat each element in a sequence k times."""
+    return [item for item in seq for _ in range(k)]
+
+
 def ray_get_with_progress(ray_refs: List[ray.ObjectRef], desc: str = "Processing", enable: bool = True) -> List[Any]:
     """Execute ray.get() with a progress bar using futures.
